Use pydantic 2 in classes
